Defining process count interval alarm
This change is to create a new alarm definition which triggers the alarm
if the number of processes is not included into a specific range defined
by default as 1-32.
Some services need to have more than one processes to run on the same
box at the same time but the number of them must not exceed a specific
bound, this new alarm can monitor this scenario.